Remove frontend and backend objects together

See merge request compsci308_2019spring/voogasalad_electricvoogaloo!69

+++ src/gameengine/src/state/action/movement/MovementAction.java
@@ -12,7 +12,7 @@ import java.util.Map;
  */
 public abstract class MovementAction extends Action implements IRequiresBaseAgent {
 
-   protected int speed;
+   protected double speed;
    protected Agent baseAgent;
 
    public MovementAction(Map<String, Object> params) {
